+++ rpmdb/rpmhash.h
@@ -48,7 +48,7 @@ int hashEqualityString(const void * key1, const void * key2);
  * @param eq            function to compare hash keys for equality
  * @return		pointer to initialized hash table
  */
-hashTable htCreate(int numBuckets, int keySize, int freeData,
+hashTable htCreate(int numBuckets, size_t keySize, int freeData,
 		hashFunctionType fn, hashEqualityType eq); 
 
 /**
